Factorize pagination comman elements addition method in pagination object
