+++ Repository/Pagination/AbstractPaginationRepository.php
@@ -31,36 +31,7 @@ abstract class AbstractPaginationRepository extends EntityRepository
         return $qb;
     }
 
-    /**
-     * @param QueryBuilder $qb
-     * @param int $page
-     * @param int $maxPerPage
-     * @return $this
-     */
-    public function addPaginationElements(QueryBuilder $qb, int $page, int $maxPerPage) {
-        return $qb
-            ->setFirstResult(($page-1) * $maxPerPage)
-            ->setMaxResults($maxPerPage)
-            ;
-    }
-
-    /**
-     * @param QueryBuilder $qb
-     * @param int $page
-     * @param int $maxPerPage
-     * @return Paginator
-     */
-    public function getListPaginated(QueryBuilder $qb, int $page, int $maxPerPage) {
-        $qb
-            ->setFirstResult(($page-1) * $maxPerPage)
-            ->setMaxResults($maxPerPage)
-        ;
-
-        $paginator = new Paginator($qb);
-        return $paginator;
-    }
-
-
+    
     /**
      * @param string $alias
      * @param array $parameters
